requirements.txt
streamlit
streamlit-authenticator
pandas
numpy
matplotlib
seaborn
plotly
google-genai
pydantic
PyMuPDF
